:recycle: Split adapter by category
